backend/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
pydantic>=2.0.0

# AI and ML
groq>=0.4.0
openai>=1.0.0

# Azure services
azure-cognitiveservices-speech>=1.30.0
azure-identity>=1.10.0

# Animation and video processing (simplified)
moviepy>=1.0.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.30.0
aiofiles>=23.0.0
httpx>=0.24.0

# Development
pytest>=7.0.0
